Adjustments in getting the db

+++ webapi/serve_webapi.py
@@ -332,27 +332,28 @@ def sendDBCopy():
 # Get The DB
 
 def getConlluDBFile():
-    while True:
-        try:
-            if VERBOSE:
-                print >> sys.stderr , "Getting the conllu file for the DB."
-
-            conn = getConnection()
-            # Download an object and save it
-            obj = conn.get_object("bosque-UD", os.path.basename(DB_CONLLU_FILE))
-            with open(DB_CONLLU_FILE, 'w') as DBfile:
-                DBfile.write(obj[1])
-
-            if VERBOSE:
-                print >> sys.stderr, "Database file %s downloaded successfully." % (DB_CONLLU_FILE)
-            break
-
-        except:
-            print >> sys.stderr,"Error getting DB file:"
-            print >> sys.stderr,traceback.format_exc()
-            print >> sys.stderr, "Trying getting the DB again..."
-            time.sleep(5)
-            continue
+    if os.path.isfile(DB_CONLLU_FILE):
+        while True:
+            try:
+                if VERBOSE:
+                    print >> sys.stderr , "Getting the conllu file for the DB."
+    
+                conn = getConnection()
+                # Download an object and save it
+                obj = conn.get_object("bosque-UD", os.path.basename(DB_CONLLU_FILE))
+                with open(DB_CONLLU_FILE, 'w') as DBfile:
+                    DBfile.write(obj[1])
+    
+                if VERBOSE:
+                    print >> sys.stderr, "Database file %s downloaded successfully." % (DB_CONLLU_FILE)
+                break
+    
+            except:
+                print >> sys.stderr,"Error getting DB file:"
+                print >> sys.stderr,traceback.format_exc()
+                print >> sys.stderr, "Trying getting the DB again..."
+                time.sleep(5)
+                continue
 
 def buildDB():
     while True:
